refactor: split paths module
